backend/pyproject.toml
[project]
name = "logistics-adk-agent"
version = "0.4.0"
description = "A Google ADK multi-agent system for cross-border logistics operations"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Candidate Submission" }]
dependencies = [
  "google-adk>=1.18.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  # --- Backend (FastAPI) ---
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.0.0",
  "python-multipart>=0.0.9",
  "pydantic[email]>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["agent*", "app*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
strict = false
